backend/requirements.txt
Django==3.2.12
djangorestframework
django-filter
psycopg2-binary
Pillow
markdown
django-cors-headers
ipython
djangorestframework-simplejwt